@@ -15,13 +15,13 @@ android {
             storePassword 'HelloWorld98!'
         }
     }
-    compileSdkVersion 27
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.blackboxindia.PostIt"
         minSdkVersion 21
-        targetSdkVersion 27
-        versionName '2.33'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 29
+        versionName '2.4'
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         versionCode 7
         manifestPlaceholders = [onesignal_app_id: "0286c78e-5fb1-4d9e-8c8e-404d81317a35",
                                 onesignal_google_project_number: "REMOTE"]
@@ -45,20 +45,20 @@ def firebase = '16.0.3'
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
 
-    implementation "com.android.support:appcompat-v7:${support}"
-    implementation "com.android.support:design:${support}"
-    implementation "com.android.support:cardview-v7:${support}"
-    implementation "com.android.support:support-v4:${support}"
-    implementation "com.android.support:percent:${support}"
-    implementation "com.android.support.constraint:constraint-layout:1.1.2"
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.percentlayout:percentlayout:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
-    implementation "com.google.firebase:firebase-core:16.0.1"
-    implementation "com.google.firebase:firebase-database:16.0.1"
-    implementation "com.google.firebase:firebase-storage:16.0.1"
-    implementation "com.google.firebase:firebase-messaging:17.3.0"
-    implementation "com.google.firebase:firebase-auth:16.0.3"
+    implementation "com.google.firebase:firebase-core:17.4.0"
+    implementation "com.google.firebase:firebase-database:19.3.0"
+    implementation "com.google.firebase:firebase-storage:19.1.1"
+    implementation "com.google.firebase:firebase-messaging:20.1.6"
+    implementation "com.google.firebase:firebase-auth:19.3.1"
 
-    implementation 'com.github.bumptech.glide:glide:4.7.1'
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
     implementation('com.github.bumptech.glide:recyclerview-integration:4.7.1') {
         transitive = false
     }
@@ -69,9 +69,11 @@ dependencies {
 
 //    implementation 'com.onesignal:OneSignal:3.7.1'
 
-    testImplementation 'junit:junit:4.12'
-    annotationProcessor "com.github.bumptech.glide:compiler:4.7.1"
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:3.0.2', {
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    //testImplementation 'junit:junit:4.12'
+
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 
