@@ -1,8 +1,8 @@
 package com.blackboxindia.PostIT.CustomViews;
 
 import android.content.Context;
-import android.support.v7.widget.CardView;
 import android.util.AttributeSet;
+import androidx.cardview.widget.CardView;
 
 
 public class SquareCard extends CardView {
