@@ -4,16 +4,14 @@ import android.app.Fragment;
 import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import com.blackboxindia.PostIT.Network.CloudStorageMethods;
 import com.blackboxindia.PostIT.Network.ConnectionDetector;
 import com.blackboxindia.PostIT.Network.Interfaces.onCompleteListener;
@@ -22,8 +20,8 @@ import com.blackboxindia.PostIT.R;
 import com.blackboxindia.PostIT.activities.MainActivity;
 import com.blackboxindia.PostIT.adapters.DocumentAdapter;
 import com.blackboxindia.PostIT.dataModels.Directory;
+import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.auth.FirebaseAuth;
-
 import io.paperdb.Paper;
 
 public class Frag_Docs extends Fragment {
