@@ -1,23 +1,24 @@
 package com.blackboxindia.PostIT.Fragments;
 
-        import android.app.AlertDialog;
-        import android.app.Fragment;
+import static com.blackboxindia.PostIT.activities.MainActivity.MY_ADS_TAG;
+import static com.blackboxindia.PostIT.activities.MainActivity.TITLE_ViewAd;
+
+import android.app.AlertDialog;
+import android.app.Fragment;
 import android.app.ProgressDialog;
 import android.content.Context;
-        import android.content.DialogInterface;
-        import android.graphics.Bitmap;
+import android.content.DialogInterface;
+import android.graphics.Bitmap;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-        import android.util.Log;
-        import android.view.LayoutInflater;
+import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import com.blackboxindia.PostIT.Network.Interfaces.onDeleteListener;
 import com.blackboxindia.PostIT.Network.NetworkMethods;
 import com.blackboxindia.PostIT.R;
@@ -25,13 +26,10 @@ import com.blackboxindia.PostIT.activities.MainActivity;
 import com.blackboxindia.PostIT.adapters.ViewAdImageAdapter;
 import com.blackboxindia.PostIT.dataModels.AdData;
 import com.blackboxindia.PostIT.dataModels.UserInfo;
+import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.storage.FileDownloadTask;
 import com.google.firebase.storage.FirebaseStorage;
-
 import java.util.List;
-
-import static com.blackboxindia.PostIT.activities.MainActivity.MY_ADS_TAG;
-import static com.blackboxindia.PostIT.activities.MainActivity.TITLE_ViewAd;
 
 public class Frag_ViewMyAd extends Fragment {
 
