@@ -1,15 +1,16 @@
 package com.blackboxindia.PostIT.Fragments;
 
+import static com.blackboxindia.PostIT.Fragments.Frag_Ads.ARGS_AdType;
+import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_LOSTFOUND;
+import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_SELL;
+import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_TEACH;
+
 import android.app.Fragment;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
@@ -19,7 +20,10 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import com.blackboxindia.PostIT.Network.Interfaces.onCompleteListener;
 import com.blackboxindia.PostIT.Network.NetworkMethods;
 import com.blackboxindia.PostIT.R;
@@ -29,14 +33,8 @@ import com.blackboxindia.PostIT.cameraIntentHelper.ImageUtils;
 import com.blackboxindia.PostIT.dataModels.AdData;
 import com.blackboxindia.PostIT.dataModels.DateObject;
 import com.blackboxindia.PostIT.dataModels.UserInfo;
-
 import java.util.ArrayList;
 import java.util.Calendar;
-
-import static com.blackboxindia.PostIT.Fragments.Frag_Ads.ARGS_AdType;
-import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_LOSTFOUND;
-import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_SELL;
-import static com.blackboxindia.PostIT.dataModels.AdTypes.TYPE_TEACH;
 
 public class Frag_newAd extends Fragment {
 
