@@ -2,15 +2,13 @@ package com.blackboxindia.PostIT.adapters;
 
 import android.content.Context;
 import android.graphics.Bitmap;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
 import android.widget.ImageView;
-
+import androidx.recyclerview.widget.RecyclerView;
 import com.blackboxindia.PostIT.R;
-
 import java.util.ArrayList;
 
 public class NewAdImageAdapter extends RecyclerView.Adapter<NewAdImageAdapter.imgViewHolder> {
